Fix the ruby field's repeated "evolution failed" messages
`UpdateFieldStructures_RubyField` (bank 8) tried to call
`LoadTimerGraphics` (bank 5) without bank switching. This invalid call
eventually lead to the PC jumping into the middle of
`CheckIfEvolutionModeTimerExpired_RedField`, which set a bunch of
indicator lights to On (since the jump was to a point after the `xor a`
instruction which would otherwise result in the indicator lights being
set to Off) and queued the "evolution failed" message.

`UpdateFieldStructures_GoldField` is also not in the same bank as
`LoadTimerGraphics`, so that call was also changed to include the bank
switch, even though there are no obvious effects from that call being
wrong.