Implement lotad bumper position changing
The bumper positions change is activated by hitting a button located between
the evolution cave and the bumpers

This button also acts as staryu in special modes

`wd7ad` means the state of the structures in RedField being a fraction of
`wStageCollisionState`; `wd7ad` was reused for RubyField to mean the state
of the lotad bumpers, to separate it from wStageCollisionState which also
includes whether the ball is currently being launched.

Remove any vestigal cases where `wd7ad` was set to `4` or `6` at the end
of special modes in RubyBoard.

Fix the indicators in the second phase of the Map Move mode since I noticed
that while looking for the `wd7ad` changes